+++ platform/flowglad-next/src/app/store/products/page.tsx
@@ -1,5 +1,0 @@
-import Internal from './Internal'
-
-export default function ProductsPage() {
-  return <Internal products={[]} />
-}